Move Operator to a different file, Add TrimPrefixName container method, Refactor ReadFile to GetFirstLine
